improved the architecture to place checks in validation instead of domain
